route handler properly return temp error
